edit: change main page profile as component
