+++ assets/scss/colors.scss
@@ -1,0 +1,3 @@
+$lightest-color: #e8e9ec;
+$darkest-color: #020406;
+$main-color: #152540;
